Create eng to chi transliteration.py
English-Chinese Transliteration
Transliteration is the conversion of a text from one script to another. For example, the Arabic script
كتا بي) i.e., written) can be converted to the Latin script kitabi for English. English words have been
also been commonly used in a various international language as loan/borrowed words such as for
technical terms (e.g., ‘computer’) or names of places (e.g., ‘New York’). Although different
languages may have different pronunciations and sound inventories, some phonetic equivalents
can be applied to a target language written in the script of the target language. For example in
Japanese, the word computer can be written as konpyutaa while in Chinese, the word tiramisu
can be written as tilamisu . Each Chinese character has a distinct meaning.

The following tasks are done :
i) a python FST program for the Chinese-Eng transliteration for loan/borrowed words
ii) an output file that prints the mappings of the transliterations as shown in Figure 1
iii) an FST construction generated by the Python program – Tkinter (image or word
document)